Handle inability to locate minicpan directory more gracefully.
This commit attempts to respond to the test failures reported by CPAN tester
Nigel Horne in report
http://www.cpantesters.org/cpan/report/1dcba6de-c545-11e4-95dd-f9a587ef3ddc.
In the test files where new() is called, the error message was "Directory
/home/njh/mirrors/minicpan not found."  This implies that a .minicpanrc was
being located but that it either lacked a 'local' setting or the directory
specified in 'local' was wrong or could not be located.  In this case, don't
proceed further with testing in 002, 005, 006, 007 and 008.