visit(): Internal modification; pass absolute path to subroutine reference.
Heretofore, we went through some contortions to isolate the basename of a
tarball from its absolute path.  But it appears this was not really necessary
to get a valid first argument to an action passed by code reference.  And
since we store distros internally by absolute path anyway, we don't need to go
through those contortions.